Angle is now a float in angle_to_transmit union
git-svn-id: https://svn.fe.up.pt/repos/lcom1920-t6g02@221 812d746c-baee-4303-96b4-7e37b0e189f4

+++ proj/src/serial_port.h
@@ -31,10 +31,10 @@ enum last_read {
 };
 
 /**
- * Help to transform the angle double into bytes to transmit and the bytes to double when receiving.
+ * Help to transform the angle float into bytes to transmit and the bytes to float when receiving.
  */
 union angle_to_transmit {
-  double angle;
+  float angle;
   uint32_t transmit;
 };
